Correct namespaces and reorganise
