+++ src/Tasks/AlgoliaReIndexTask.php
@@ -6,12 +6,16 @@ use SilverStripe\Dev\BuildTask;
 use SilverStripe\Dev\Debug;
 use SilverStripe\SiteConfig\SiteConfig;
 use SilverStripe\Control\HTTPRequest;
-
+use Chewyou\Algolia\Services\AlgoliaIndexService;
 
 class AlgoliaReIndexTask extends BuildTask
 {
+    private static $segment = 'AlgoliaReIndexTask';
+
     public function run($request)
     {
-        Debug::dump("Success");
+        $service = new AlgoliaIndexService();
+
+        return $service->run();
     }
 }
